# docs/openapi.yaml
openapi: 3.0.3
info:
  title: cardinal-wings API
  description: >
    REST API daemon for managing cardinal over HTTP. Panel-oriented surface
    over cardinal's Docker-compatible API, with multi-node routing (`?node=`),
    async tasks, streaming logs/exec, metrics and role-based auth.
  version: "0.2.0-dev"
servers:
  - url: http://127.0.0.1:8080
security:
  - bearerAuth: []
...
paths:
...
  /v1/containers:
    get:
      tags: [containers]
      summary: List containers
      parameters:
        - { name: all, in: query, schema: { type: boolean } }
        - { name: state, in: query, schema: { type: string, enum: [running, stopped] } }
        - { name: image, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string } }
        - { name: sort, in: query, schema: { type: string, enum: [name, status, created_at] } }
        - { name: order, in: query, schema: { type: string, enum: [asc, desc] } }
        - { name: limit, in: query, schema: { type: integer, default: 100 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
        - { name: node, in: query, schema: { type: string } }
      responses:
        "200":
          description: Container page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContainerPage" }
    post:
      tags: [containers]
      summary: Create a container (admin)
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRequest" }
      responses:
        "201": { description: Created }
  /v1/containers/{id}:
    get:
      tags: [containers]
      summary: Inspect a container
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: node, in: query, schema: { type: string } }
      responses:
        "200": { description: Container detail }
    delete:
      tags: [containers]
      summary: Remove a container (admin)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: force, in: query, schema: { type: boolean } }
      responses:
        "200": { description: Removed }
...
  /v1/containers/{id}/exec/stream:
    post:
      tags: [containers]
      summary: Run a command and stream output as SSE (admin)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExecRequest" }
      responses: { "200": { description: SSE stream } }
...
  /v1/containers/{id}/terminal/input:
    post:
      tags: [containers]
      summary: Write to the terminal session stdin (admin)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                data: { type: string }
      responses: { "200": { description: ok } }
  /v1/containers/{id}/terminal/resize:
    post:
      tags: [containers]
      summary: Record the panel viewport for the session (admin, 1..1000)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [cols, rows]
              properties:
                cols: { type: integer }
                rows: { type: integer }
      responses: { "200": { description: session with cols and rows } }
  /v1/containers/{id}/terminal/stream:
    get:
      tags: [containers]
      summary: Read the terminal session output (SSE)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: SSE stream } }
  /v1/containers/{id}/terminal/ws:
    get:
      tags: [containers]
      summary: Terminal session over websocket (admin)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: websocket stream } }
...
  /v1/containers/{id}/fm/list:
    get:
      tags: [containers]
      summary: File manager list (admin reads allowed)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: path, in: query, schema: { type: string } }
      responses: { "200": { description: entries } }
  /v1/containers/{id}/fm/read:
    get:
      tags: [containers]
      summary: File manager read file
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: path, in: query, schema: { type: string } }
      responses: { "200": { description: content } }
  /v1/containers/{id}/fm/download:
    get:
      tags: [containers]
      summary: File manager download file (base64)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: path, in: query, schema: { type: string } }
      responses: { "200": { description: base64 content } }
...
  /v1/containers/{id}/ports/add:
    post:
      tags: [containers]
      summary: Add a port mapping without restart (admin)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                mapping: { type: string }
      responses: { "200": { description: added } }
  /v1/containers/{id}/ports/remove:
    post:
      tags: [containers]
      summary: Remove a port mapping (admin)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                mapping: { type: string }
      responses: { "200": { description: removed } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ContainerPage:
      type: object
      properties:
        containers: { type: array }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
    CreateRequest:
      type: object
      required: [image]
      properties:
        image: { type: string }
        name: { type: string }
        ports: { type: array, items: { type: string } }
        env: { type: array, items: { type: string } }
        volumes: { type: array, items: { type: string } }
        restart: { type: string }
        memory: { type: string }
        cpus: { type: number }
        labels: { type: object, additionalProperties: { type: string } }
    ExecRequest:
      type: object
      required: [Cmd]
      properties:
        Cmd: { type: array, items: { type: string } }
        AttachStdin: { type: boolean }
        AttachStdout: { type: boolean }
        AttachStderr: { type: boolean }
        Tty: { type: boolean }
...